Move DnsResolverCompat
