+++ core/src/main/java/com/github/shadowsocks/net/DnsResolverCompat.kt
@@ -18,7 +18,7 @@
  *                                                                             *
  *******************************************************************************/
 
-package com.github.shadowsocks.bg
+package com.github.shadowsocks.net
 
 import android.annotation.SuppressLint
 import android.annotation.TargetApi
